requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-filter==22.1
drf-spectacular==0.24.2